+++ src/main/java/pl/koder95/bso/service/impl/CategoryServiceImpl.java
@@ -52,6 +52,10 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public void deleteById(Long id) {
-        categoryRepository.deleteById(id);
+        if (categoryRepository.existsById(id)) {
+            categoryRepository.deleteById(id);
+        } else {
+            throw new EntityNotFoundException("Cannot find a category by id: " + id);
+        }
     }
 }
